+++ CHANGELOG.md
@@ -10,6 +10,7 @@ Henad uses [semantic versioning](https://semver.org/spec/v2.0.0.html) for its re
 - Model metadata in the Model tab in GUI
 - Export tab in the GUI app
 - Unlimited history setting in the Charts tab
+- Model actions (see `henad_core::actions!`)
 
 ## [0.1.0] - 2026-09-08